オセロのルールに関するreducerを別クラスに切り出した
